feat: stepRounds 改为 Long 类型，支持更大滚动步长范围
将 byIdRule.stepRounds 从 Integer 改为 Long 类型，完整覆盖从前端到后端到 DB 整条链路：
- 后端实体 ArchiveGroupItemById.stepRounds: Integer → Long
- SyncExecutor 新增 execute(Long, Long, int, Long) 重载
- DbArchiveRuleLoader 使用 rs.getLong 读取 DB 结果
- Request/VO 模型 stepRounds: Integer → Long
- Service 校验新增 validatePositive(Long) 重载
- DB schema step_rounds: INT → BIGINT
- 前端 TypeScript stepRounds: number → string（避免 JS Long 精度丢失）
- 前端表单 v-model.number → v-model，默认值 5000 → "5000"

+++ easyarchive-core/src/main/java/com/openquartz/easyarchive/core/rule/DbArchiveRuleLoader.java
@@ -91,7 +91,7 @@ public class DbArchiveRuleLoader implements ArchiveRuleLoader {
         rule.setStartId(rs.getString("start_id"));
         rule.setEndId(rs.getString("end_id"));
         rule.setStepCount(rs.getInt("step_count"));
-        rule.setStepRounds(rs.getInt("step_rounds"));
+        rule.setStepRounds(rs.getLong("step_rounds"));
 
         return rule;
     }
